requirements.txt
faiss-cpu==1.7.4
langchain==0.3.26
langchain-community==0.3.27
langchain-huggingface==0.3.0
huggingface-hub==0.34.4
sentence-transformers
python-dotenv
transformers
torch
streamlit